requirements.txt
anthropic
python-telegram-bot==21.6
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv